fixed: Compiler warning
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@9545 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/libPython/xbmcmodule/controllist.cpp
@@ -506,7 +506,7 @@ PyDoc_STRVAR(addItem__doc__,
     int iPos = -1;
     if (!PyArg_ParseTuple(args, "i", &iPos)) return NULL;
 
-    if (iPos < 0 || iPos >= self->vecItems.size())
+    if (iPos < 0 || iPos >= (int)self->vecItems.size())
     {
       PyErr_SetString(PyExc_ValueError, "Index out of range");
       return NULL;
